NMR Number of Scans Estimator

adjustr() is the keystone function that the package is based on. It takes the input values of what an NMR spectrum's signal and noise values are and how many scans it took to produce them and returns the number of scans need to be performed by an NMR to produce the desired signal to noise ratio as defined in the goal arg.

adjustr(sig, noi, goal, ns, r2)

Arguments

sig
The signal

noi
The noise

goal
The target signal to noise

ns
The number of scans used to create the provided signal and noise values

r2
Arguments passed to embedded roundr function. See roundr help for details.

Value

The number of scans (rounded based on the input arguments from r2) necessary to achieve the goal signal to noise

Details

The signal to noise ratio of an NMR spectrum shows how large the signal is compared to that of the noise. Adjusting the number of done by an NMR spectrometer can allow you to alter this signal to noise ratio. The relationship between number of scans and signal to noise ratio is based around a square root so if one wants to increase the the signal to noise ratio by a factor of two, they must increase the number of scans by a factor of 4. adjustr() takes this relationship into account to calculate the number of scans needed to be performed on a sample to produce a spectrum with a certain signal to noise and is the value returned by the function.

Author

Kyle Bario kylebario1@gmail.com

Examples

data(m)
ns <- adjustr(300000, 1500, 40000, m$a_NS, c(4,512,2))